Remediate recovery review findings for issue #4.
Reset recovery counters on successful toolUse turns without early continuation, split index runtime helpers under 1000 lines, and make retry tests deterministic via injectable delay scheduling and fake timers.
