+++ src/index.ts
@@ -3,13 +3,20 @@ import type { ExtensionAPI, ExtensionContext } from "@earendil-works/pi-coding-a
 import { registerGoalCommand } from "./commands.js";
 import { formatFooterStatus } from "./format.js";
 import {
-  budgetLimitPrompt,
-  compactContinuationPrompt,
-  continuationGoalIdFromPrompt,
-  continuationPrompt,
-  supersededContinuationMessage,
-} from "./prompts.js";
-import { applyUsage, clearEntry, goalWithLiveUsage, goalsEquivalent, reconstructGoal, setEntry, updateGoalStatus } from "./state.js";
+  assistantTurnTokens,
+  createAccountingState,
+  createGoalAccounting,
+  isAbortedAssistantMessage,
+  isToolUseAssistantMessage,
+} from "./goal-accounting.js";
+import { compactContinuationPrompt, continuationGoalIdFromPrompt } from "./prompts.js";
+import {
+  dedupeActiveGoalContinuations,
+  extensionQueuedGoalWorkMessageId,
+  extensionQueuedGoalWorkMessageIdForRuntime,
+  pendingStaleQueuedGoalWorkIdsFromMessages,
+  staleGoalContinuationContextMessage,
+} from "./queued-goal-work.js";
 import {
   bumpRecoveryGeneration,
   createGoalRecoveryMachine,
@@ -18,287 +25,12 @@ import {
 } from "./recovery-machine.js";
 import { createGoalRecoveryRuntime } from "./recovery-runtime.js";
 import { isErrorAssistantMessage } from "./recovery.js";
+import { clearEntry, goalWithLiveUsage, goalsEquivalent, reconstructGoal, setEntry, updateGoalStatus } from "./state.js";
 import { registerGoalTools } from "./tools.js";
 import { CUSTOM_ENTRY_TYPE, type GoalEntrySource, type GoalResult, type ThreadGoal } from "./types.js";
 
-interface AccountingState {
-  activeGoalId: string | null;
-  lastAccountedAt: number | null;
-  budgetWarningSentFor: string | null;
-}
-
 interface StatusContext {
   ui: Pick<ExtensionContext["ui"], "setStatus">;
-}
-
-interface AssistantUsage {
-  input: number;
-  output: number;
-}
-
-interface AssistantTurnMessage {
-  role: string;
-  stopReason?: string;
-  errorMessage?: string;
-  usage?: AssistantUsage;
-}
-
-interface QueuedGoalMessageDetails {
-  kind?: unknown;
-  goalId?: unknown;
-}
-
-interface TextMessagePart {
-  type?: unknown;
-  text?: unknown;
-}
-
-function usageChannelTokens(value: number): number {
-  if (!Number.isFinite(value)) {
-    return 0;
-  }
-  return Math.max(0, Math.trunc(value));
-}
-
-function assistantTurnTokens(message: AssistantTurnMessage): number {
-  if (message.role !== "assistant" || !message.usage) {
-    return 0;
-  }
-  return usageChannelTokens(message.usage.input) + usageChannelTokens(message.usage.output);
-}
-
-function isAbortedAssistantMessage(message: AssistantTurnMessage): boolean {
-  return message.role === "assistant" && message.stopReason === "aborted";
-}
-
-function isToolUseAssistantMessage(message: AssistantTurnMessage): boolean {
-  return message.role === "assistant" && message.stopReason === "toolUse";
-}
-
-function isQueuedGoalWorkKind(kind: unknown): boolean {
-  return kind === "continuation" || kind === "command_start" || kind === "command_resume";
-}
-
-function isSupersededContinuationDetails(details: unknown): boolean {
-  return isQueuedGoalMessageDetails(details) && details.kind === "superseded_continuation";
-}
-
-function isQueuedGoalMessageDetails(details: unknown): details is QueuedGoalMessageDetails {
-  return details !== null && typeof details === "object";
-}
-
-function textContentFromMessageContent(content: unknown): string | null {
-  if (typeof content === "string") {
-    return content;
-  }
-
-  if (!Array.isArray(content)) {
-    return null;
-  }
-
-  const textParts: string[] = [];
-  for (const part of content) {
-    if (part === null || typeof part !== "object") {
-      continue;
-    }
-    const textPart = part as TextMessagePart;
-    if (textPart.type === "text" && typeof textPart.text === "string") {
-      textParts.push(textPart.text);
-    }
-  }
-
-  return textParts.length > 0 ? textParts.join("\n") : null;
-}
-
-function continuationGoalIdFromMessageContent(content: unknown): string | null {
-  const text = textContentFromMessageContent(content);
-  return text === null ? null : continuationGoalIdFromPrompt(text);
-}
-
-function staleGoalContinuationMessage(queuedGoalId: string, currentGoal: ThreadGoal | null): string {
-  const currentState = currentGoal
-    ? `Current goal id: ${currentGoal.goalId}; current status: ${currentGoal.status}.`
-    : "There is no current goal.";
-  return [
-    "A queued hidden goal continuation was stale and has been cancelled before running.",
-    `Queued goal id: ${queuedGoalId}.`,
-    currentState,
-    "Ignore only this stale hidden bookkeeping message; do not perform work for the queued goal id above or mention this cancellation to the user.",
-  ].join("\n");
-}
-
-function extensionQueuedGoalWorkMessageId(message: {
-  role: string;
-  customType?: string;
-  details?: unknown;
-  content?: unknown;
-}): string | null {
-  if (message.role !== "custom" || message.customType !== CUSTOM_ENTRY_TYPE) {
-    return null;
-  }
-
-  if (isSupersededContinuationDetails(message.details)) {
-    return null;
-  }
-
-  if (isQueuedGoalMessageDetails(message.details)) {
-    const { kind, goalId } = message.details;
-    if (isQueuedGoalWorkKind(kind) && typeof goalId === "string") {
-      return goalId;
-    }
-  }
-
-  return continuationGoalIdFromMessageContent(message.content);
-}
-
-function queuedGoalWorkMessageId(message: {
-  role: string;
-  customType?: string;
-  details?: unknown;
-  content?: unknown;
-}): string | null {
-  if (message.role === "user") {
-    return continuationGoalIdFromMessageContent(message.content);
-  }
-
-  return extensionQueuedGoalWorkMessageId(message);
-}
-
-function supersededContinuationContextMessage<TMessage extends { role: string; content?: unknown; display?: boolean; details?: unknown }>(
-  message: TMessage,
-  goalId: string,
-): TMessage {
-  const content = supersededContinuationMessage(goalId);
-
-  if (message.role === "custom") {
-    return {
-      ...message,
-      content,
-      display: false,
-      details: {
-        kind: "superseded_continuation",
-        goalId,
-      },
-    } as TMessage;
-  }
-
-  return {
-    ...message,
-    content: [{ type: "text", text: content }],
-  } as TMessage;
-}
-
-function continuationPromptForProviderContext(
-  goal: ThreadGoal,
-  message: { details?: unknown },
-): string {
-  if (isQueuedGoalMessageDetails(message.details)) {
-    const kind = message.details.kind;
-    if (kind === "command_start" || kind === "command_resume") {
-      return continuationPrompt(goal);
-    }
-  }
-
-  return compactContinuationPrompt(goal);
-}
-
-function dedupeActiveGoalContinuations<TMessage extends {
-  role: string;
-  customType?: string;
-  details?: unknown;
-  content?: unknown;
-  display?: boolean;
-}>(
-  messages: TMessage[],
-  goal: ThreadGoal,
-  resolveQueuedGoalWorkMessageId: (
-    message: { role: string; customType?: string; details?: unknown; content?: unknown },
-  ) => string | null,
-): { messages: TMessage[]; changed: boolean } {
-  const activeGoalId = goal.goalId;
-  const indices: number[] = [];
-  for (let index = 0; index < messages.length; index += 1) {
-    const message = messages[index];
-    if (!message) {
-      continue;
-    }
-    const queuedGoalId = resolveQueuedGoalWorkMessageId(message);
-    if (queuedGoalId === activeGoalId) {
-      indices.push(index);
-    }
-  }
-
-  const latestIndex = indices.at(-1);
-  if (latestIndex === undefined) {
-    return { messages, changed: false };
-  }
-
-  let changed = false;
-  const nextMessages = messages.slice();
-
-  for (const index of indices.slice(0, -1)) {
-    const message = nextMessages[index];
-    if (!message) {
-      continue;
-    }
-    nextMessages[index] = supersededContinuationContextMessage(message, activeGoalId);
-    changed = true;
-  }
-
-  const latestMessage = nextMessages[latestIndex];
-  if (!latestMessage) {
-    return { messages, changed };
-  }
-  const refreshedContent = continuationPromptForProviderContext(goal, latestMessage);
-  if (latestMessage.role === "custom") {
-    if (latestMessage.content !== refreshedContent) {
-      nextMessages[latestIndex] = {
-        ...latestMessage,
-        content: refreshedContent,
-        display: false,
-      };
-      changed = true;
-    }
-  } else {
-    const refreshedUserContent = [{ type: "text", text: refreshedContent }];
-    const currentContent = textContentFromMessageContent(latestMessage.content);
-    if (currentContent !== refreshedContent) {
-      nextMessages[latestIndex] = {
-        ...latestMessage,
-        content: refreshedUserContent,
-      } as TMessage;
-      changed = true;
-    }
-  }
-
-  return { messages: nextMessages, changed };
-}
-
-function staleGoalContinuationContextMessage<TMessage extends { role: string; content?: unknown }>(
-  message: TMessage,
-  queuedGoalId: string,
-  currentGoal: ThreadGoal | null,
-): TMessage {
-  const content = staleGoalContinuationMessage(queuedGoalId, currentGoal);
-
-  if (message.role === "custom") {
-    return {
-      ...message,
-      content,
-      display: false,
-      details: {
-        kind: "stale_continuation",
-        goalId: queuedGoalId,
-        currentGoalId: currentGoal?.goalId ?? null,
-        currentStatus: currentGoal?.status ?? null,
-      },
-    } as TMessage;
-  }
-
-  return {
-    ...message,
-    content: [{ type: "text", text: content }],
-  } as TMessage;
 }
 
 const CONTINUATION_RETRY_MS = 50;
@@ -321,11 +53,7 @@ export default function (pi: ExtensionAPI): void {
   let startedStaleQueuedGoalWorkThisTurn = false;
   let startedRunnableWorkThisTurn = false;
   const startedStaleQueuedGoalWorkGoalIds = new Set<string>();
-  const accounting: AccountingState = {
-    activeGoalId: null,
-    lastAccountedAt: null,
-    budgetWarningSentFor: null,
-  };
+  const accounting = createAccountingState();
   let recoveryState: GoalRecoveryMachineState = createGoalRecoveryMachine();
   let errorRecoveryTimer: ReturnType<typeof setTimeout> | null = null;
 
@@ -498,27 +226,8 @@ export default function (pi: ExtensionAPI): void {
     customType?: string;
     details?: unknown;
     content?: unknown;
-  }): string | null => {
-    if (message.role === "user") {
-      const text = textContentFromMessageContent(message.content);
-      return text === null ? null : continuationGoalIdFromRuntimePrompt(text);
-    }
-
-    return queuedGoalWorkMessageId(message);
-  };
-
-  const pendingStaleQueuedGoalWorkIdsFromMessages = (
-    messages: Array<{ role: string; customType?: string; details?: unknown; content?: unknown }>,
-  ): string[] => {
-    const goalIds: string[] = [];
-    for (const message of messages) {
-      const queuedGoalId = queuedGoalWorkMessageId(message);
-      if (queuedGoalId !== null && staleQueuedGoalWorkAgentEndGoalIds.has(queuedGoalId)) {
-        goalIds.push(queuedGoalId);
-      }
-    }
-    return goalIds;
-  };
+  }): string | null =>
+    extensionQueuedGoalWorkMessageIdForRuntime(message, continuationGoalIdFromRuntimePrompt);
 
   const skipStaleQueuedGoalWorkTurnEnd = (
     turnIndex: number | null,
@@ -560,7 +269,7 @@ export default function (pi: ExtensionAPI): void {
       return false;
     }
 
-    const staleGoalIds = pendingStaleQueuedGoalWorkIdsFromMessages(messages);
+    const staleGoalIds = pendingStaleQueuedGoalWorkIdsFromMessages(messages, staleQueuedGoalWorkAgentEndGoalIds);
     if (staleGoalIds.length === 0) {
       return false;
     }
@@ -644,60 +353,16 @@ export default function (pi: ExtensionAPI): void {
     refreshUi(ctx);
   };
 
-  const beginAccounting = (): void => {
-    if (!goal || goal.status !== "active") {
-      accounting.activeGoalId = null;
-      accounting.lastAccountedAt = null;
-      return;
-    }
-
-    accounting.activeGoalId = goal.goalId;
-    accounting.lastAccountedAt = Date.now();
-  };
-
-  const accountProgress = (
-    ctx: ExtensionContext,
-    allowBudgetSteering: boolean,
-    completedTurnTokens = 0,
-    accountBudgetLimited = false,
-  ): void => {
-    const canAccount = goal?.status === "active" || (accountBudgetLimited && goal?.status === "budgetLimited");
-    if (!goal || accounting.activeGoalId !== goal.goalId || !canAccount) {
-      beginAccounting();
-      return;
-    }
-
-    const now = Date.now();
-    const elapsed = accounting.lastAccountedAt === null ? 0 : Math.floor((now - accounting.lastAccountedAt) / 1000);
-    accounting.lastAccountedAt = now;
-
-    const result = applyUsage(goal, completedTurnTokens, elapsed, {
-      expectedGoalId: accounting.activeGoalId,
-      accountBudgetLimited,
-    });
-    if (!result.changed || !result.goal) {
-      return;
-    }
-
-    persistGoal(result.goal, "runtime");
-    refreshUi(ctx);
-
-    if (allowBudgetSteering && result.crossedBudget && accounting.budgetWarningSentFor !== result.goal.goalId) {
-      accounting.budgetWarningSentFor = result.goal.goalId;
-      pi.sendMessage(
-        {
-          customType: CUSTOM_ENTRY_TYPE,
-          content: budgetLimitPrompt(result.goal),
-          display: false,
-          details: { kind: "budget_limit", goalId: result.goal.goalId },
-        },
-        { triggerTurn: true, deliverAs: "steer" },
-      );
-    }
-  };
+  const goalAccounting = createGoalAccounting({
+    getGoal: () => goal,
+    getAccounting: () => accounting,
+    persistGoal,
+    refreshUi,
+    sendMessage: pi.sendMessage.bind(pi),
+  });
 
   const completeGoal = (source: GoalEntrySource, ctx: ExtensionContext): GoalResult => {
-    accountProgress(ctx, false, 0, true);
+    goalAccounting.accountProgress(ctx, false, 0, true);
     const result = updateGoalStatus(goal, "complete");
     if (!result.ok || !result.goal) {
       return result;
@@ -851,7 +516,7 @@ export default function (pi: ExtensionAPI): void {
         noteStaleQueuedGoalWorkTerminalEvents();
       }
       staleQueuedGoalWorkTurnActive = true;
-      clearActiveAccounting();
+      goalAccounting.clearActiveAccounting();
       ctx.abort();
       refreshUi(ctx);
     }
@@ -861,12 +526,12 @@ export default function (pi: ExtensionAPI): void {
 
   pi.on("session_start", async (event, ctx) => {
     reloadFromSession(ctx);
-    beginAccounting();
+    goalAccounting.beginAccounting();
     if (event.reason === "resume" && goal?.status === "paused" && ctx.hasUI) {
       const shouldResume = await ctx.ui.confirm("Resume paused goal?", `Goal: ${goal.objective}`);
       if (shouldResume) {
         resumePausedGoal(ctx);
-        beginAccounting();
+        goalAccounting.beginAccounting();
       }
     }
     maybeContinue(ctx);
@@ -874,7 +539,7 @@ export default function (pi: ExtensionAPI): void {
 
   pi.on("session_tree", async (_event, ctx) => {
     reloadFromSession(ctx);
-    beginAccounting();
+    goalAccounting.beginAccounting();
     maybeContinue(ctx);
   });
 
@@ -919,7 +584,7 @@ export default function (pi: ExtensionAPI): void {
     recoveryRuntime.clearHandledErrorTurn();
     clearStartedTurnWork();
     clearStaleQueuedGoalWorkTurn();
-    beginAccounting();
+    goalAccounting.beginAccounting();
     refreshUi(ctx);
   });
 
@@ -928,7 +593,7 @@ export default function (pi: ExtensionAPI): void {
       return;
     }
 
-    accountProgress(ctx, true, 0, true);
+    goalAccounting.accountProgress(ctx, true, 0, true);
   });
 
   pi.on("turn_end", async (_event, ctx) => {
@@ -937,7 +602,7 @@ export default function (pi: ExtensionAPI): void {
     }
 
     const completedTurnTokens = assistantTurnTokens(_event.message);
-    accountProgress(ctx, true, completedTurnTokens);
+    goalAccounting.accountProgress(ctx, true, completedTurnTokens);
     if (isAbortedAssistantMessage(_event.message)) {
       pauseForAbort(ctx);
       return;
@@ -946,9 +611,9 @@ export default function (pi: ExtensionAPI): void {
       recoveryRuntime.handleAssistantError(_event.message, ctx, _event.turnIndex);
       return;
     }
-    if (!isToolUseAssistantMessage(_event.message)) {
-      recoveryRuntime.finishSuccessfulAssistantTurn(_event.message, ctx);
-    }
+    recoveryRuntime.finishSuccessfulAssistantTurn(_event.message, ctx, {
+      continueGoal: !isToolUseAssistantMessage(_event.message),
+    });
   });
 
   pi.on("agent_end", async (event, ctx) => {
@@ -961,7 +626,7 @@ export default function (pi: ExtensionAPI): void {
     const abortedTurnTokens = abortedMessages.reduce((sum, message) => {
       return sum + assistantTurnTokens(message);
     }, 0);
-    accountProgress(ctx, false, abortedTurnTokens, true);
+    goalAccounting.accountProgress(ctx, false, abortedTurnTokens, true);
     if (abortedMessages.length > 0) {
       pauseForAbort(ctx);
       return;
@@ -983,7 +648,7 @@ export default function (pi: ExtensionAPI): void {
       return;
     }
 
-    accountProgress(ctx, false, 0, true);
+    goalAccounting.accountProgress(ctx, false, 0, true);
   });
 
   pi.on("session_compact", async (_event, ctx) => {
@@ -1010,7 +675,7 @@ export default function (pi: ExtensionAPI): void {
     }
     clearStaleQueuedGoalWorkTerminalEvents();
 
-    accountProgress(ctx, false, 0, true);
+    goalAccounting.accountProgress(ctx, false, 0, true);
     clearContinuationTimer();
     clearErrorRecoveryTimer();
     stopStatusRefresh();
