+++ test/index.test.ts
@@ -1,5 +1,5 @@
 import assert from "node:assert/strict";
-import test from "node:test";
+import test, { mock } from "node:test";
 
 import type { ExtensionAPI, ExtensionCommandContext, ExtensionContext } from "@earendil-works/pi-coding-agent";
 
@@ -10,7 +10,7 @@ import {
   continuationGoalIdFromPrompt,
   continuationPrompt,
 } from "../src/prompts.js";
-import { recoveryAttentionMessage } from "../src/recovery.js";
+import { recoveryAttentionMessage, transientErrorBackoffMs } from "../src/recovery.js";
 import { isGoalCustomEntry, reconstructGoal } from "../src/state.js";
 import { CUSTOM_ENTRY_TYPE } from "../src/types.js";
 
@@ -1916,78 +1916,88 @@ test("compaction failures pause with recoverable attention", async () => {
 });
 
 test("repeated transient errors use bounded backoff before pausing", async () => {
-  const harness = createRuntimeHarness();
-  await harness.runCommand("ship it");
-  harness.sentMessages.length = 0;
+  mock.timers.enable({ apis: ["setTimeout"] });
+  try {
+    const harness = createRuntimeHarness();
+    await harness.runCommand("ship it");
+    harness.sentMessages.length = 0;
 
-  for (let attempt = 0; attempt < 6; attempt += 1) {
-    await harness.emit("turn_start", { type: "turn_start", turnIndex: attempt, timestamp: attempt + 1 });
-    await harness.emit("turn_end", {
-      type: "turn_end",
-      turnIndex: attempt,
-      message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
-      toolResults: [],
-    });
-    if (attempt < 5) {
-      await new Promise((resolve) => setTimeout(resolve, 1_100 * 2 ** attempt));
+    for (let attempt = 0; attempt < 6; attempt += 1) {
+      await harness.emit("turn_start", { type: "turn_start", turnIndex: attempt, timestamp: attempt + 1 });
+      await harness.emit("turn_end", {
+        type: "turn_end",
+        turnIndex: attempt,
+        message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
+        toolResults: [],
+      });
+      if (attempt < 5) {
+        mock.timers.tick(transientErrorBackoffMs(attempt + 1));
+      }
     }
-  }
 
-  assert.equal(harness.snapshot().goal?.status, "paused");
-  assert.equal(harness.sentMessages.length, 0);
+    assert.equal(harness.snapshot().goal?.status, "paused");
+    assert.equal(harness.sentMessages.length, 0);
+  } finally {
+    mock.timers.reset();
+  }
 });
 
 test("successful turns reset transient error counters and continue active goals", async () => {
-  const harness = createRuntimeHarness();
-  await harness.runCommand("ship it");
-  const queued = harness.sentMessages[0];
-  assert.ok(queued);
-  const queuedMessage = queuedCustomMessage(queued);
-  harness.sentMessages.length = 0;
+  mock.timers.enable({ apis: ["setTimeout"] });
+  try {
+    const harness = createRuntimeHarness();
+    await harness.runCommand("ship it");
+    const queued = harness.sentMessages[0];
+    assert.ok(queued);
+    const queuedMessage = queuedCustomMessage(queued);
+    harness.sentMessages.length = 0;
 
-  await harness.emit("turn_start", { type: "turn_start", turnIndex: 0, timestamp: 1 });
-  await harness.emit("message_start", {
-    type: "message_start",
-    message: queuedMessage,
-  });
-  await harness.emit("turn_end", {
-    type: "turn_end",
-    turnIndex: 0,
-    message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
-    toolResults: [],
-  });
-  assert.equal(harness.sentMessages.length, 0);
+    await harness.emit("turn_start", { type: "turn_start", turnIndex: 0, timestamp: 1 });
+    await harness.emit("message_start", {
+      type: "message_start",
+      message: queuedMessage,
+    });
+    await harness.emit("turn_end", {
+      type: "turn_end",
+      turnIndex: 0,
+      message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
+      toolResults: [],
+    });
+    assert.equal(harness.sentMessages.length, 0);
 
-  await new Promise((resolve) => setTimeout(resolve, 1_100));
-  await harness.emit("turn_start", { type: "turn_start", turnIndex: 1, timestamp: 2 });
-  await harness.emit("turn_end", {
-    type: "turn_end",
-    turnIndex: 1,
-    message: assistantMessage("stop", { input: 1, output: 1 }),
-    toolResults: [],
-  });
+    mock.timers.tick(transientErrorBackoffMs(1));
+    await harness.emit("turn_start", { type: "turn_start", turnIndex: 1, timestamp: 2 });
+    await harness.emit("turn_end", {
+      type: "turn_end",
+      turnIndex: 1,
+      message: assistantMessage("stop", { input: 1, output: 1 }),
+      toolResults: [],
+    });
 
-  assert.equal(harness.snapshot().goal?.status, "active");
-  assert.equal(harness.sentMessages.length, 1);
+    assert.equal(harness.snapshot().goal?.status, "active");
+    assert.equal(harness.sentMessages.length, 1);
 
-  harness.sentMessages.length = 0;
-  await harness.emit("before_agent_start", {
-    type: "before_agent_start",
-    prompt: "keep going",
-    systemPrompt: "",
-    systemPromptOptions: {},
-  });
-  await harness.emit("turn_start", { type: "turn_start", turnIndex: 2, timestamp: 3 });
-  await harness.emit("turn_end", {
-    type: "turn_end",
-    turnIndex: 2,
-    message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
-    toolResults: [],
-  });
-  await new Promise((resolve) => setTimeout(resolve, 1_100));
+    harness.sentMessages.length = 0;
+    await harness.emit("before_agent_start", {
+      type: "before_agent_start",
+      prompt: "keep going",
+      systemPrompt: "",
+      systemPromptOptions: {},
+    });
+    await harness.emit("turn_start", { type: "turn_start", turnIndex: 2, timestamp: 3 });
+    await harness.emit("turn_end", {
+      type: "turn_end",
+      turnIndex: 2,
+      message: assistantMessage("error", { input: 1, output: 1 }, "websocket closed"),
+      toolResults: [],
+    });
+    mock.timers.tick(transientErrorBackoffMs(1));
 
-  assert.equal(harness.snapshot().goal?.status, "active");
-  assert.equal(harness.sentMessages.length, 1);
+    assert.equal(harness.snapshot().goal?.status, "active");
+    assert.equal(harness.sentMessages.length, 1);
+  } finally {
+    mock.timers.reset();
+  }
 });
 
 test("unavailable compaction pauses with recoverable attention", async () => {
@@ -2052,26 +2062,71 @@ test("exhausted context overflow retries show recoverable attention in footer", 
 });
 
 test("turn_end and agent_end dedupe a single assistant error into one retry", async () => {
+  mock.timers.enable({ apis: ["setTimeout"] });
+  try {
+    const harness = createRuntimeHarness();
+    await harness.runCommand("ship it");
+    harness.sentMessages.length = 0;
+
+    const errorMessage = assistantMessage("error", { input: 1, output: 1 }, "websocket closed");
+    await harness.emit("turn_start", { type: "turn_start", turnIndex: 0, timestamp: 1 });
+    await harness.emit("turn_end", {
+      type: "turn_end",
+      turnIndex: 0,
+      message: errorMessage,
+      toolResults: [],
+    });
+    await harness.emit("agent_end", {
+      type: "agent_end",
+      messages: [errorMessage],
+    });
+
+    mock.timers.tick(transientErrorBackoffMs(1));
+    assert.equal(harness.snapshot().goal?.status, "active");
+    assert.equal(harness.sentMessages.length, 0);
+  } finally {
+    mock.timers.reset();
+  }
+});
+
+test("successful toolUse turns reset context overflow recovery counters", async () => {
   const harness = createRuntimeHarness();
   await harness.runCommand("ship it");
   harness.sentMessages.length = 0;
 
-  const errorMessage = assistantMessage("error", { input: 1, output: 1 }, "websocket closed");
   await harness.emit("turn_start", { type: "turn_start", turnIndex: 0, timestamp: 1 });
   await harness.emit("turn_end", {
     type: "turn_end",
     turnIndex: 0,
-    message: errorMessage,
+    message: assistantMessage("error", { input: 1, output: 1 }, "context_length_exceeded"),
     toolResults: [],
   });
-  await harness.emit("agent_end", {
-    type: "agent_end",
-    messages: [errorMessage],
+  assert.equal(harness.compactCalls.length, 1);
+  await harness.emit("session_compact", {
+    type: "session_compact",
+    summary: "compact summary",
+    tokensBefore: 100,
   });
 
-  await new Promise((resolve) => setTimeout(resolve, 1_100));
-  assert.equal(harness.snapshot().goal?.status, "active");
+  await harness.emit("turn_start", { type: "turn_start", turnIndex: 1, timestamp: 2 });
+  await harness.emit("turn_end", {
+    type: "turn_end",
+    turnIndex: 1,
+    message: assistantMessage("toolUse", { input: 1, output: 1 }),
+    toolResults: [],
+  });
   assert.equal(harness.sentMessages.length, 0);
+
+  await harness.emit("turn_start", { type: "turn_start", turnIndex: 2, timestamp: 3 });
+  await harness.emit("turn_end", {
+    type: "turn_end",
+    turnIndex: 2,
+    message: assistantMessage("error", { input: 1, output: 1 }, "context_length_exceeded"),
+    toolResults: [],
+  });
+
+  assert.equal(harness.compactCalls.length, 2);
+  assert.equal(harness.snapshot().goal?.status, "active");
 });
 
 test("stale compaction callbacks do not pause a replaced active goal", async () => {
